backend/requirements.txt
fastapi==0.120.3
uvicorn[standard]
starlette==0.49.1
pydantic
openai==2.6.1
python-dotenv==1.2.1
requests==2.32.5
httpx
ipykernel==7.1.0
